♻️ [desktop]: Move quick pick window functions to windows module
Consolidate window management by moving quick pick window creation
and positioning logic from quick_pick.rs to windows.rs.
